INTERNAL: Move the pallet argument processor to its rightful place
