+++ common/src/stack/command/stack/commands/enable/pallet/__init__.py
@@ -13,10 +13,11 @@
 
 import os
 import stack.commands
+from stack.argument_processors.pallet import PalletArgumentProcessor
 from stack.exception import ArgRequired, CommandError
 
 
-class Command(stack.commands.PalletArgumentProcessor,
+class Command(PalletArgumentProcessor,
 	stack.commands.enable.command):
 	"""
 	Enable an available pallet. The pallet must already be copied on the
@@ -91,7 +92,7 @@ class Command(stack.commands.PalletArgumentProcessor,
 		# Remember the box info to simplify queries down below
 		box_id, box_os = rows[0]
 
-		for pallet in self.getPallets(args, params):
+		for pallet in self.get_pallets(args, params):
 			# Make sure this pallet's OS is valid for the box
 			if box_os != pallet.os:
 				raise CommandError(self,
