+++ common/src/stack/command/stack/commands/list/pallet/tag/__init__.py
@@ -6,9 +6,10 @@
 
 import fnmatch
 import stack.commands
+from stack.argument_processors.pallet import PalletArgumentProcessor
 
 class Command(stack.commands.list.command,
-	      stack.commands.PalletArgumentProcessor):
+	      PalletArgumentProcessor):
 	"""
 	Lists the set of tags for hosts.
 
@@ -16,7 +17,7 @@ class Command(stack.commands.list.command,
 	Name of the pallet. If no pallet is specified tags are listed
 	for all pallets.
 	</arg>
-	
+
 	<param type='string' name='tag' optional='0'>
 	A shell syntax glob pattern to specify to tags to
 	be listed.
@@ -41,9 +42,9 @@ class Command(stack.commands.list.command,
 
 	def run(self, params, args):
 
-		pallets = self.getPallets(args, params)
+		pallets = self.get_pallets(args, params)
 
-		(glob, ) = self.fillParams([ 
+		(glob, ) = self.fillParams([
 			('tag', '*')
 			])
 
@@ -58,7 +59,7 @@ class Command(stack.commands.list.command,
 				tags[pid] = {}
 			if fnmatch.fnmatch(tname, glob):
 				tags[pid][tname] = tvalue
-		
+
 		self.beginOutput()
 
 		for pallet in pallets:
@@ -66,12 +67,12 @@ class Command(stack.commands.list.command,
 				continue
 			for key in sorted(tags[pallet.id].keys()):
 				value = tags[pallet.id][key]
-				self.addOutput(pallet.name, 
-					       (pallet.version, pallet.rel, 
-						pallet.arch, pallet.os, 
+				self.addOutput(pallet.name,
+					       (pallet.version, pallet.rel,
+						pallet.arch, pallet.os,
 					        key, value))
 
-		self.endOutput(('pallet', 'version', 'release', 'arch', 'os', 
+		self.endOutput(('pallet', 'version', 'release', 'arch', 'os',
 				'tag', 'value'))
